+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.2.0 (????-??-??) ##
 
 - Power ranking projections for unfinished leagues.
+- Caches fantasy content for each user for up to 6 hours (configurable).
 
 ## 0.1.0 (2014-03-02) ##